+++ moderngl_window/__init__.py
@@ -16,8 +16,9 @@ from moderngl_window.context.base import WindowConfig, BaseWindow
 from moderngl_window.timers.clock import Timer
 from moderngl_window.conf import settings
 from moderngl_window.utils.module_loading import import_string
+from moderngl_window.utils.keymaps import KeyMapFactory, KeyMap, QWERTY, AZERTY  # noqa
 
-__version__ = "2.3.0"
+__version__ = "2.4.4"
 
 IGNORE_DIRS = [
     "__pycache__",
@@ -203,6 +204,7 @@ def run_window_config(config_cls: WindowConfig, timer=None, args=None) -> None:
         vsync=values.vsync if values.vsync is not None else config_cls.vsync,
         samples=values.samples if values.samples is not None else config_cls.samples,
         cursor=show_cursor if show_cursor is not None else True,
+        backend=values.backend,
     )
     window.print_context_info()
     activate_context(window=window)
@@ -290,6 +292,10 @@ def create_parser():
         type=valid_window_size_multiplier,
         default=1.0,
         help="Multiplier for the window size making it easy scale the window",
+    )
+    parser.add_argument(
+        "--backend",
+        help="Specify context backend. This is mostly used to enable EGL in headless mode",
     )
     return parser