+++ moderngl_window/integrations/imgui.py
@@ -123,11 +123,11 @@ class ModernGLRenderer(BaseOpenGLRenderer):
         if not self.ctx:
             raise ValueError("Missing moderngl context")
 
-        assert isinstance(self.ctx, moderngl.context.Context)
+        assert isinstance(self.ctx, moderngl.Context)
 
         super().__init__()
 
-        if hasattr(self, "wnd"):
+        if hasattr(self, "wnd") and self.wnd:
             self.resize(*self.wnd.buffer_size)
         elif "display_size" in kwargs:
             self.io.display_size = kwargs.get("display_size")